pdf-QA-chatbot-main/pdf-QA-chatbot-main/requirement.txt
langchain==0.0.154
PyPdf2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2 
langchain-community
google-generativeai
InstructorEmbedding
transformers
torch
